Fix sign mismatch warning.

+++ source/soccer/assignment/cost_gk.cpp
@@ -5,6 +5,6 @@ namespace Tyr::Soccer
 int Ai::gkRoleCost(const int t_robot_idx, const Assignment &t_assignment) const
 {
     (void) t_assignment;
-    return m_ref_state.ourInfo().gk_id == t_robot_idx ? 0 : -1;
+    return static_cast<int>(m_ref_state.ourInfo().gk_id) == t_robot_idx ? 0 : -1;
 }
 } // namespace Tyr::Soccer
